requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.0.4
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django_vite_plugin==3.0.0
djangorestframework==3.15.1
djlint==1.34.1
EditorConfig==0.12.4
environs==11.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
jsbeautifier==1.15.1
json5==0.9.25
Markdown==3.6
marshmallow==3.21.1
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.2
urllib3==2.2.1
uvicorn==0.29.0
